OY2-20 added optional attachments (#26)
Updated uploads component styling to matched agreed styling from 9/17/2020 meeting.
Added optional uploads.
Updated jsdoc with more information.
Updated email template with upload titles.
Minor fixes to build tools.
Made the failure error message shown to the user a bit less techie.
Made some changes to get to an offline run with a local s3, but further refactoring will be needed. Right now we can start the offline s3 bucket, but the UI code needs to be changed to use it, which is complicated due to the amplify library not allowing different endpoints.