Oy2 175 cms access attachments (#32)
Allow permanent read access to attachments. Before this change, we could only obtain links that would expire after a max of 7 days. This was a problem as CMS needs access to the attachments past the 7 days. This change open public access to the attachments.
Switched the attachment upload to protected instead of private for a possible future with CMS receivers logging in to see the files.