OY2 915 (subtask of OY2-108) Added SPA RAI Form (#35)
Added SPA RAI Submission form
Changed the create backend to be generic. We check for required fields and the rest are provided by the UI.
Left create.js for now, to limit merge issues, but it is not used.
SPA RAI using CMS Design System
Refactored list of records to be generic.
Other minor cleanups.