OY2-1526 Updated waiver keys and cleanups (#60)
Updated the waiver action types and authority keys per issue.
Added submitted at date and state for records in database. The point was to be able to see what records the CMS email was sent and which the email fails without having to add code to delete the record if there was an error. This may be useful for CQRS. Note that we will really never see the created state unless SES is completely down.
Some minor code cleanups.
Improvements to configureLocal, so errors can be seen.