OY2-2635 validate attachment size (#85)
Fixed the validation of attachment sizes.
Changes the configurable parameter to be in MB instead of bytes.
Removed the attachment max size hint and error to be only when creating a request. It was showing in the view only as well.
Ran the formatter on the files.
Removed duplicate serverless config in app-api
Refactored File Uploader to handle the error messages. Two possible errors can be shown at a time.