OY2 8778 Form Page bug fixes (#268)
Merging this even with Nightwatch tests failing because the test needs to be updated and it is not a trivial update.  The test incorrectly uses the file upload input DOM object as the place where the file list is kept and the application keeps that list in state and allows manipulation of the list that isn't also managed by the inputs.