fix for carets being changed in UI
something happened in a PR that made the automation unable to interact with them. One bit is that the count for carets increased by 1-2K and even using first still gave that many results. Also, the code started treating the caret like it was hidden. All tests are green when ran locally so this should work!

Also new data added so uncommented out a test