added/update readme to point to confluence
ideally we would have the full documentation here in github and link to the github documetion from confluence. However the story indicated the other way around and also there is no installed markdown app in confluence that can pull from github so it would make the confluence page a straight link.