Updated status map to also account for substatus
While I think having the status map include the substatuses is fine, it would also be prudent given a longer time horizon to rebaseline the actual status as well. However not a big deal since that is the purpose of the map, you can have  a different backend status vs what shows to the user.